refactor: shift env func to utils, simplify config funcs
