+++ auto_update.py
@@ -8,7 +8,7 @@ from bittensor.utils.btlogging import logging as logger
 
 from commons.utils import datetime_to_iso8601_str
 from dojo import get_latest_git_tag
-from dojo.utils.config import source_dotenv
+from dojo.utils import source_dotenv
 
 source_dotenv()
