+++ dojo/logging/__init__.py
@@ -1,0 +1,3 @@
+from .format import apply_custom_logging_format
+
+__all__ = ["apply_custom_logging_format"]
