+++ dojo/logging/format.py
