+++ main_miner.py
@@ -4,7 +4,7 @@ from bittensor.utils.btlogging import logging as logger
 
 from commons.block_subscriber import start_block_subscriber
 from commons.objects import ObjectManager
-from dojo.utils.config import source_dotenv
+from dojo.utils import source_dotenv
 
 source_dotenv()
