+++ scripts/extract_dataset.py
@@ -20,7 +20,7 @@ from database.prisma.types import (
     ValidatorTaskWhereInput,
 )
 from dojo.protocol import Scores
-from dojo.utils.config import source_dotenv
+from dojo.utils import source_dotenv
 
 source_dotenv()
